refactor(simulation): separate delete delivery from history
